% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysumfunction.R
\name{mysumfunction}
\alias{mysumfunction}
\title{Title}
\usage{
mysumfunction(s)
}
\arguments{
\item{param}{\itemize{
\item A numeric value
}}
}
\value{
Output the input number plus by 1
}
\description{
Title
}
\examples{
mysumfunction(1)

}
